+++ examples/workflows/mapreduce/runner/main.go
@@ -1,0 +1,32 @@
+package main
+
+import (
+	"context"
+	"log/slog"
+
+	"github.com/tilebox/tilebox-go/examples/workflows/mapreduce"
+	"github.com/tilebox/tilebox-go/workflows/v1"
+)
+
+func main() {
+	ctx := context.Background()
+	client := workflows.NewClient()
+
+	runner, err := client.NewTaskRunner(ctx)
+	if err != nil {
+		slog.Error("failed to create task runner", slog.Any("error", err))
+		return
+	}
+
+	err = runner.RegisterTasks(
+		&mapreduce.RootTask{},
+		&mapreduce.MapTask{},
+		&mapreduce.ReduceTask{},
+	)
+	if err != nil {
+		slog.Error("failed to register tasks", slog.Any("error", err))
+		return
+	}
+
+	runner.RunForever(ctx)
+}