+++ workflows/v1/job/options.go
@@ -31,10 +31,31 @@ func WithClusterSlug(clusterSlug string) SubmitOption {
 	}
 }
 
+// State is an alias to the workflows State type for use in query options.
+type State int32
+
+// State values for use in query filters.
+const (
+	// Submitted means the job has been submitted, queued and waiting for its first task to be run.
+	Submitted State = 1
+	// Running means the job is running, i.e. at least one task is running.
+	Running State = 2
+	// Started means the job has started running, i.e. at least one task has been computed, but currently no tasks are running.
+	Started State = 3
+	// Completed means the job has completed successfully.
+	Completed State = 4
+	// Failed means the job has failed.
+	Failed State = 5
+	// Canceled means the job has been canceled on user request.
+	Canceled State = 6
+)
+
 // QueryOptions contains the configuration for a Query request.
 type QueryOptions struct {
 	TemporalExtent query.TemporalExtent
-	AutomationID   uuid.UUID
+	AutomationIDs  []uuid.UUID
+	States         []State
+	Name           string
 }
 
 type QueryOption func(*QueryOptions)
@@ -47,9 +68,25 @@ func WithTemporalExtent(temporalExtent query.TemporalExtent) QueryOption {
 	}
 }
 
-// WithAutomationID specifies the automation ID to filter jobs by.
-func WithAutomationID(automationID uuid.UUID) QueryOption {
+// WithAutomationIDs specifies multiple automation IDs to filter jobs by.
+// Only jobs submitted by any of the specified automations will be returned.
+func WithAutomationIDs(automationIDs ...uuid.UUID) QueryOption {
 	return func(cfg *QueryOptions) {
-		cfg.AutomationID = automationID
+		cfg.AutomationIDs = append(cfg.AutomationIDs, automationIDs...)
+	}
+}
+
+// WithJobStates filters jobs by their state.
+// Only jobs in any of the specified states will be returned.
+func WithJobStates(states ...State) QueryOption {
+	return func(cfg *QueryOptions) {
+		cfg.States = append(cfg.States, states...)
+	}
+}
+
+// WithJobName filters jobs by name.
+func WithJobName(name string) QueryOption {
+	return func(cfg *QueryOptions) {
+		cfg.Name = name
 	}
 }